pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "betx"
version = "1.0.0"
description = "Système d'analyse de paris sportifs multi-sports avec value betting"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.25",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "statsmodels>=0.14",
    "requests>=2.31",
    "sqlalchemy>=2.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "streamlit>=1.28",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "httpx>=0.25",
    "apscheduler>=3.10",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
include = ["betx*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
